requirements.txt
psutil
mysql-connector-python
dash
pandas
plotly
